#8 Database Schema Setup
Using Hibernate and JPA I defined the User class using annotatins to specify table names, create indexes on frequently accessed data. Hibernate maps the table to the class and with JPA's help it creates he DB and table automatically.

Also defined a simple controller to check DB connection and that everything is working