Appointment Service and Availability Service
Created a business logic layer for the appointments fetching data from the other APIs and enriching the appoitment details.

Created a simple availability service with fixed 30min intervals jus for the MVP later on this will be based on services durations.

Included a security config to protect endpoints, enable cors configuration to access other services and JWT decoder for fetching role from the jwt and enable role-based access to certain endpoints